requirements.txt
streamlit
pandas
matplotlib
azure-ai-textanalytics
azure-core
python-dotenv